% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc.redistribute-package.R
\docType{package}
\name{sdc.redistribute-package}
\alias{sdc.redistribute}
\alias{sdc.redistribute-package}
\title{sdc.redistribute: Redistribute Values Between Geographic Areas}
\description{
Estimate attribute values for one set of polygons from values measured on a different, misaligned set. Provides area-weighted areal interpolation and a dasymetric method that distributes values across a point layer (such as parcel centroids). Count (extensive) measures are total-preserving; rate (intensive) measures use area-weighted means.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dads2busy/sdc.redistribute}
  \item Report bugs at \url{https://github.com/dads2busy/sdc.redistribute/issues}
}

}
\author{
\strong{Maintainer}: Aaron Schroeder \email{ads7fg@virginia.edu} (\href{https://orcid.org/0000-0001-9985-1136}{ORCID})

}
\keyword{internal}
